Move _looks_like_financial_year into helpers
Co-locates it with the other financial-year utilities so detect and
convert live together.
